Antes de cambio a seleccionar workstation en monitor

+++ Inventario5/Models/Monitor.cs
@@ -8,7 +8,7 @@ namespace Inventario5.Models
     public partial class Monitor
     {
         public int Id { get; set; }
-        public byte? TamanoPantalla { get; set; }
+        public float? TamanoPantalla { get; set; }
         public int? ResolucionX { get; set; }
         public int? ResolucionY { get; set; }
         public int? IdWorkstation { get; set; }
